% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Title: calculate_distance}
\usage{
calculate_distance(XStart, XEnd, YStart, YEnd)
}
\arguments{
\item{XStart}{vector of starting longitude points}

\item{XEnd}{vector of ending longitude points}

\item{YStart}{vector of starting latitude points}

\item{YEnd}{vector of ending latitude points}
}
\value{
A vector with path lengths in kms
}
\description{
Calculates the geodetic distance between two points
}
