% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_locations.R
\docType{data}
\name{sample_locations}
\alias{sample_locations}
\title{Sample coordinates of microwave links}
\format{
Vector of 4 vectors (XStart, XEnd, YStart, Yend)
}
\source{
Randomly generated in R
}
\usage{
sample_locations
}
\description{
Coordinates to draw 100 lines;de generated between 79.9 and 81.8 degrees lon. and 6 and 9.5 lat.
}
\keyword{datasets}
